+++ src/documents.jl
@@ -314,6 +314,7 @@ struct User
     linkcheck_timeout::Real   # ..but only wait this many seconds for each one.
     linkcheck_useragent::String  # User agent to use for linkchecks.
     checkdocs::Symbol         # Check objects missing from `@docs` blocks. `:none`, `:exports`, or `:all`.
+    checkdocs_ignored_modules::Vector{Module}  # ..and then ignore (some of) them.
     doctestfilters::Vector{Regex} # Filtering for doctests
     warnonly::Vector{Symbol}  # List of docerror groups that should only warn, rather than cause a build failure
     pages   :: Vector{Any}    # Ordering of document pages specified by the user.
@@ -388,6 +389,7 @@ function Document(;
         linkcheck_timeout :: Real    = 10,
         linkcheck_useragent :: String= _LINKCHECK_DEFAULT_USERAGENT,
         checkdocs::Symbol            = :all,
+        checkdocs_ignored_modules::Vector{Module} = Module[],
         doctestfilters::Vector{Regex}= Regex[],
         warnonly :: Union{Bool,Symbol,Vector{Symbol}} = Symbol[],
         modules  :: Union{Module, Vector{Module}} = Module[],
@@ -454,6 +456,7 @@ function Document(;
         linkcheck_timeout,
         linkcheck_useragent,
         checkdocs,
+        checkdocs_ignored_modules,
         doctestfilters,
         warnonly,
         pages,
@@ -494,7 +497,7 @@ function Document(;
 
     blueprint = DocumentBlueprint(
         Dict{String, Page}(),
-        submodules(modules),
+        submodules(modules; ignore=Set(checkdocs_ignored_modules)),
     )
     Document(user, internal, plugin_dict, blueprint)
 end