+++ src/utilities/utilities.jl
@@ -201,19 +201,19 @@ end
 """
 Returns the set of submodules of a given root module/s.
 """
-function submodules(modules::Vector{Module})
+function submodules(modules::Vector{Module}; ignore = Set{Module}())
     out = Set{Module}()
     for each in modules
-        submodules(each, out)
+        submodules(each, out; ignore=ignore)
     end
     out
 end
-function submodules(root::Module, seen = Set{Module}())
+function submodules(root::Module, seen = Set{Module}(); ignore = Set{Module}())
     push!(seen, root)
     for name in names(root, all=true)
         if Base.isidentifier(name) && isdefined(root, name) && !isdeprecated(root, name)
             object = getfield(root, name)
-            if isa(object, Module) && !(object in seen) && parentmodule(object::Module) == root
+            if isa(object, Module) && !(object in seen) && !(object in ignore) && parentmodule(object::Module) == root
                 submodules(object, seen)
             end
         end