+++ test/missingdocs/make.jl
@@ -13,6 +13,15 @@ module MissingDocs
     g(x) = x
 end
 
+module MissingDocsSubmodule
+    module UndocumentedSubmodule
+        export f
+
+        "exported"
+        f(x) = x
+    end
+end
+
 @testset "missing docs" begin
     for (sym, n_expected) in zip([:none, :exports, :all], [0, 1, 2])
         kwargs = (
@@ -38,6 +47,23 @@ end
         checkdocs = :all,
         sitename = "MissingDocs Checks",
     )
+
+    for (ignore, n_expected) in zip([false, true] , [1, 0])
+        kwargs = (
+            root = dirname(@__FILE__),
+            source = joinpath("src", "none"),
+            build = joinpath("build", "submodule"),
+            modules = MissingDocsSubmodule,
+            checkdocs = :all,
+            sitename = "MissingDocsSubmodule Checks",
+            warnonly = true,
+            checkdocs_ignored_modules = ignore ? Module[MissingDocsSubmodule.UndocumentedSubmodule] : Module[],
+        )
+        @quietly @test makedocs(; kwargs...) === nothing
+
+        doc = Documenter.Document(; kwargs...)
+        @quietly @test Documenter.missingdocs(doc) == n_expected
+    end
 end
 
 end